# FIX-#2742: fix performance degradation for dictionary GroupBy aggregation