@amyskov 2021-02-18
Failed to find functional tests for case with multicolumn groupby and dictionary aggregation (are they really present?). If not, it will be good to extend test suite with this case.