@dchigarev 2021-02-18
@amyskov [`test_multi_column_groupby_different_partitions`](https://github.com/modin-project/modin/blob/8ebbad9bca36aba0ddfc58dac89b9a66b0bfe41a/modin/pandas/test/test_groupby.py#L1534) tests multi-column GroupBy with basic and dictionary aggregation